re-use rows on stack computed in previous branch
